feat: add continue ui
