+++ src/api/app.py
@@ -60,6 +60,7 @@ def create_app() -> Flask:
 
     load_plugins_from_dir(app, "flaskr/service/study/input")
     load_plugins_from_dir(app, "flaskr/service/study/ui")
+    load_plugins_from_dir(app, "flaskr/service/study/continue")
 
     try:
         load_plugins_from_dir(app, "flaskr/plugins", plugin_manager)
