+++ src/api/flaskr/service/study/input/handle_input_continue.py
@@ -32,9 +32,9 @@ def handle_input_continue(
     db.session.add(log_script)
     span = trace.span(name="user_continue", input=input)
     span.end()
+    db.session.flush()
     continue_func = CONTINUE_HANDLE_MAP.get(script_info.script_ui_type, None)
     if continue_func:
-        return continue_func(
+        continue_func(
             app, user_info, lesson, attend, script_info, input, trace, trace_args
         )
-    db.session.flush()
